+++ TeaseLibFramework/JNIUtilities.h
@@ -13,7 +13,7 @@ class JNIUtilities
 public:
 	static std::vector<std::wstring> stringArray(JNIEnv* env, jobjectArray jarray);
 
-	static jobject newList(JNIEnv* env, jint capacity);
+	static jobject newList(JNIEnv* env, size_t capacity);
 
 	static jobject asList(JNIEnv* env, const std::vector<std::string>& elements);
 	static jobject asList(JNIEnv* env, const std::vector<NativeObject*>& elements);
